% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/climate_matching_create_climate_difference_matrix.r
\name{create_climate_difference_matrix}
\alias{create_climate_difference_matrix}
\title{Climate matching}
\usage{
create_climate_difference_matrix(output_file_name, target_data_percentiles,
  target_data_monthly_mean_growing_season, comparison_data_percentiles,
  comparison_data_monthly_mean_growing_season, target_coordinates = list(x =
  149.13, y = 35.2809))
}
\arguments{
\item{output_file_name}{path to output file}

\item{target_data_percentiles}{output of the function brick_calc_percentiles for the 'target' region and time period.}

\item{target_data_monthly_mean_growing_season}{output of the function of \code{transform_monthly_means_into_growing_season_means} for the 'target' region and time period.}

\item{comparison_data_percentiles}{output of the function \code{brick_calc_percentiles} for the 'comparison' region and time period.}

\item{comparison_data_monthly_mean_growing_season}{output of the function of transform_monthly_means_into_growing_season_means for the 'comparison' region and time period.}

\item{target_coordinates}{must be a named list in the form of: list(x=x_vals,y=y_vals) in coordinates that match the input data; or, the full file PATH to a shapefile}
}
\value{
Returns a file-backed raster object.  i.e. It creates a .grd file, that can also bee accessible in the environment if assigned to a variable.
}
\description{
This function is designed to compare climatic values of one variable in two different time periods from a specific region.
}
\details{
The 'target' is grid that contains the location and time period you wish to compare to everywhere, at the comparison time, whereas, the 'comparison' is the grid that represents the rest of the world at the desired time period (e.g. if you want to find out where in the world will be similar to what Hobart is now, in 2080, then the 'target' grid needs to be "the present" and the comparison grid need to be "the future").

This function requires the outputs from for both the 'target' and the 'comparison' grids:
\code{brick_calc_percentiles}: a raster brick of 99 layers, each representating the percentiles from the a specific set of ' source, variable, time period, region'.
\code{transform_monthly_means_into_growing_season_means}: a raster brick of 12 layers, each representing the monthly mean value for each month of the 'growing season'.  NB: 'growing season' monthly means is when the southern hemisphere has been offset by 6 months, so seasons 'align' globally (as much as they can be, i.e. in the Northern Hemisphere, 1=January, 12=December; in the Southern Hemisphere, 1=July, 12=June).
}
\examples{
#I need to put in an example
}
