% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/transform_monthly_means_into_growing_season_means.r
\name{transform_monthly_means_into_growing_season_means}
\alias{transform_monthly_means_into_growing_season_means}
\title{Calculate growing season monthly means}
\usage{
transform_monthly_means_into_growing_season_means(brick_of_monthly_means)
}
\arguments{
\item{brick_of_monthly_means}{output from \code{brick_calc_monthly_mean}. A raster brick of 12 layers representing the 12 months of the year.}
}
\value{
a raster brick object with 12 layers, indicating the 12 months of the year, starting in the middle of winter.
}
\description{
Gowing seasons are approximately offset by 12 months between the Northern and Southern Hemispheres.  Although this is a gross simplification, when trying to compare monthly values globally, it is necessary to align the summers and the winters, reffered to in this context as the growing seasons.  This function aims to achive this by shifting the Southern hemishere values of a given grid by 6 months.
}
\examples{
#I need to add an exmple...
}
